requirements.txt
requests
Django==6.0.3
python-dotenv
psycopg[binary]
playwright